update video
Former-commit-id: 857d93e4b2907bb7c027ea58a599c1ccf04acf80

+++ cmd/user/service/user_service.go
@@ -28,7 +28,7 @@ func (s *UserService) GetUserById(ctx context.Context, request *user.UserRequest
 		return
 	}
 	resp_user = &user.User{
-		Id:              int64(db_user.ID),
+		Id:              db_user.ID,
 		Name:            db_user.UserName,
 		Avatar:          &db_user.Avatar,
 		BackgroundImage: &db_user.BackgroundImage,
